requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
pylatexenc>=2.10
pymupdf>=1.24.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
Pillow>=10.3.0
